Switch to delta time game loop and fix collision logic
Implemented a delta time game loop in `MainPannel`.
Refactored `Entity`, `Player`, and `CollisionChecker` to use `double` for positions and `dt` for movement calculations.
Fixed a crash in `CollisionChecker` by adding boundary checks.
Fixed collision logic in `Player` to check for collisions before moving.
Supported diagonal movement and collision checking.

+++ src/Entity/Entity.java
@@ -5,12 +5,12 @@ import java.awt.Rectangle;
 
 public class Entity {
 
-    public int WorldX, WorldY;
-    public int speed;
+    public double WorldX, WorldY;
+    public double speed;
 
     public BufferedImage standing1, standing2, standing3,standing4, standing5, standing6, up1, up2, up3, up4, up5, up6, down1, down2, down3, down4, down5, down6, left1, left2, left3, left4, left5, left6, right1, right2, right3, right4, right5, right6, fight1, fight2, fight3, fight4;
     public String direction;
-    
+
 
     public int spriteCounter = 0;
     public int spriteCounter_for_idle = 0;
@@ -18,10 +18,10 @@ public class Entity {
 
     public int spriteNum = 1;
     public int spriteNum_for_standing_animation = 1;
-    public int spriteNum_for_attack = 1; 
+    public int spriteNum_for_attack = 1;
 
     public Rectangle  solidArea;
     public int solidAreadefaultX, solidAreadefaultY;
     public boolean playercollision = false;
-    
+
 }
